Rename latency -> latency1 globally
